+++ src/Network/Network/Models/AzureFirewall/PSAzureFirewall.cs
@@ -78,6 +78,8 @@ namespace Microsoft.Azure.Commands.Network.Models
 
         public List<string> Zones { get; set; }
 
+        public string AllowActiveFTP { get; set; }
+
         [JsonIgnore]
         public string IpConfigurationsText
         {